feat(router): added react router
Added reat router to the project
