+++ package.json
@@ -18,10 +18,11 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-redux": "^7.2.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "redux": "^4.0.5",
-    "typeface-roboto": "^0.0.75",
     "redux-thunk": "^2.3.0",
+    "typeface-roboto": "^0.0.75",
     "typescript": "~3.7.2"
   },
   "scripts": {
@@ -48,6 +49,7 @@
   "devDependencies": {
     "@types/axios": "^0.14.0",
     "@types/react-redux": "^7.1.9",
+    "@types/react-router-dom": "^5.1.5",
     "@types/redux-thunk": "^2.1.0"
   }
 }
