+++ src/components/Home/Home.test.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import Home from './Home';
+
+test('renders learn react link', () => {
+  const { getByText } = render(<Home />);
+  const linkElement = getByText(/learn react/i);
+  expect(linkElement).toBeInTheDocument();
+});
